Fixed error in cycle detection code
Created a new, dedicated function to detect cycles and resolved the
code. Removed a few unused constants and added some comments. Switched
the primary experiment to an oscillation between two Boolean goals as
used in the experiments conducted by Kashtan and Alon. The results do
not seem to converge to perfect solutions like previous code — this may
be because his algorithms do not work well with proper code to detect
cycles, or there may still be a pending bug that is impairing
performance.